Refactored and added enter key
