+++ app/src/main/java/com/example/klavier/KlavierApp.kt
@@ -2,14 +2,7 @@ package com.example.klavier
 
 import androidx.annotation.StringRes
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -18,7 +11,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.asLiveData
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
@@ -26,9 +18,9 @@ import androidx.navigation.compose.rememberNavController
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import com.example.klavier.data.Layout
-import com.example.klavier.ui.MainScreen
-import com.example.klavier.ui.SettingsScreen
-import com.example.klavier.ui.SplashScreen
+import com.example.klavier.ui.screens.MainScreen
+import com.example.klavier.ui.screens.SettingsScreen
+import com.example.klavier.ui.screens.SplashScreen
 
 enum class KlavierScreen(@StringRes val title: Int) {
     Start(title = R.string.splash_screen),
@@ -36,34 +28,10 @@ enum class KlavierScreen(@StringRes val title: Int) {
     Settings(title = R.string.settings)
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun KlavierAppBar(
-    currentScreen: KlavierScreen,
-    canNavigateBack: Boolean,
-    navigateUp: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    TopAppBar(
-        title = { Text(stringResource(currentScreen.title)) },
-        modifier = modifier,
-        navigationIcon = {
-            if (canNavigateBack) {
-                IconButton(onClick = navigateUp) {
-                    Icon(
-                        imageVector = Icons.Filled.Settings,//Icons.Filled.ArrowBack,
-                        contentDescription = "settings"//stringResource(R.string.back_button)
-                    )
-                }
-            }
-        }
-    )
-}
-
 @Composable
 fun KlavierApp(
     viewModel: USBViewModel,
-    SettingViewModel: SettingModelView,
+    SettingViewModel: SettingViewModel,
     modifier: Modifier = Modifier,
     navController: NavHostController = rememberNavController(),
     isDarkTheme: Boolean,
@@ -81,15 +49,7 @@ fun KlavierApp(
     var macroIcons by remember { mutableStateOf(ArrayList<Int>()) }
     var macroFunctions by remember { mutableStateOf(ArrayList<() -> Unit>()) }
 
-    Scaffold(
-       /* topBar = {
-            KlavierAppBar(
-                currentScreen = currentScreen,
-                canNavigateBack = navController.previousBackStackEntry != null,
-                navigateUp = { navController.navigateUp() }
-            )
-        }*/
-    ) { innerPadding ->
+    Scaffold{ innerPadding ->
         NavHost(
             navController = navController,
             startDestination = KlavierScreen.Start.name,
