+++ app/src/main/java/com/example/klavier/SettingViewModel.kt
@@ -1,18 +1,15 @@
 package com.example.klavier
 
 import android.content.Context
-import android.util.Log
-import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.klavier.data.Layout
 import com.example.klavier.data.SettingPreferenceRepository
 import kotlinx.coroutines.launch
 
-
-class SettingModelView(private val settingPreferencesRepository: SettingPreferenceRepository,
+class SettingViewModel(private val settingPreferencesRepository: SettingPreferenceRepository,
                        private val sendData: (String) -> Unit
-,private val context: Context
+                       , private val context: Context
 ) : ViewModel() {
 
     val settingPreferences = settingPreferencesRepository.SettingPreferencesFlow
