--- app/src/main/java/com/example/klavier/ui/MainScreen.kt
@@ -1,742 +1,0 @@
-package com.example.klavier.ui
-
-
-import android.content.Context
-import androidx.annotation.DrawableRes
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material3.Button
-import androidx.compose.material3.Card
-import androidx.compose.material3.FilterChip
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedCard
-import androidx.compose.material3.Tab
-import androidx.compose.material3.TabRow
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.TextField
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.input.key.Key
-import androidx.compose.ui.input.key.nativeKeyCode
-import androidx.compose.ui.input.key.onKeyEvent
-import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
-import com.example.klavier.R
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.github.skydoves.colorpicker.compose.*
-import kotlin.math.abs
-import androidx.compose.ui.window.Dialog
-
-@Composable
-fun MainScreen(
-    sendData: (String) -> Unit,
-    onSettingsButtonClicked : () -> Unit = {},
-    macroLabels: ArrayList<String>,
-    macroIcons: ArrayList<Int>,
-    macroFunctions: ArrayList<() -> Unit>
-)
-{
-    var tabIndex by remember { mutableIntStateOf(0) }
-    val tabs = listOf("Macros", "ColorPicker")
-
-    Column{
-        Row(
-            modifier = Modifier
-                .align(Alignment.CenterHorizontally)
-                .fillMaxWidth(),horizontalArrangement = Arrangement.Center
-        ){
-            Text(
-                "Klavier",
-                modifier = Modifier
-                    .align(Alignment.CenterVertically)
-                    .padding(16.dp),
-                style = MaterialTheme.typography.titleLarge,
-                textAlign = TextAlign.Center,
-            )
-
-            IconButton(
-                onClick = onSettingsButtonClicked,
-                Modifier.align(Alignment.CenterVertically)
-            ) {
-                Icon(
-                    imageVector = Icons.Filled.Settings,
-                    contentDescription = "Settings")
-            }
-        }
-        HorizontalDivider(
-            thickness = 1.dp, // Épaisseur de la ligne
-            color = Color.Gray // Couleur de la ligne
-        )
-        TabRow(
-            selectedTabIndex = tabIndex, containerColor = MaterialTheme.colorScheme.background, contentColor = MaterialTheme.colorScheme.onPrimary) {
-            tabs.forEachIndexed { index, title ->
-                Tab(text = { Text(title) },
-                    selected = tabIndex == index,
-                    onClick = { tabIndex = index }
-                )
-            }
-        }
-        Column (modifier  = Modifier.fillMaxSize(1f)) {
-            when (tabIndex) {
-                0 -> {
-                    MacrosTab(
-                        sendData = sendData,
-                        macroFunctions = macroFunctions,
-                        macroIcons = macroIcons,
-                        macroLabels = macroLabels
-
-                    )
-                    TouchPad(sendData = sendData)
-                }
-                1 -> ColorPickerTab(sendData = sendData)
-            }
-
-        }
-        Row (modifier = Modifier.align(Alignment.CenterHorizontally)
-        ){
-            KeyboardInput(sendData)
-        }
-    }
-}
-
-
-
-@Composable
-fun MacrosTab(
-    sendData: (String) -> Unit,
-    macroLabels: ArrayList<String>,
-    macroIcons: ArrayList<Int>,
-    macroFunctions: ArrayList<() -> Unit>
-)
-{
-    val context = LocalContext.current
-    Column{
-        LazyRow(
-            modifier = Modifier
-                .height(50.dp)
-        ) {
-            // macro copier
-            item{
-                GridItem(
-                    onClick = {
-                        sendData(context.getString(R.string.id_copy))
-                    },
-                    painter = R.drawable.copy_icon,
-                    contentDescription = "copier"
-                )
-            }
-            // macro coller
-            item{
-                GridItem(
-                    onClick = {
-                        sendData(context.getString(R.string.id_paste))
-                    },
-                    painter = R.drawable.paste_icon,
-                    contentDescription = "coller"
-                )
-            }
-            items(macroLabels) {arrayItem ->
-                GridItem(
-                    onClick = macroFunctions[macroLabels.indexOf(arrayItem)],
-                    painter = macroIcons[macroLabels.indexOf(arrayItem)],
-                    contentDescription = arrayItem
-                )
-            }
-        }
-        Row(modifier = Modifier.align(Alignment.Start))
-        {
-            var showDialog by remember { mutableStateOf(false) }
-            if (showDialog) {
-                AddMacrosDialog(onDismissRequest = { showDialog = false }, onConfirmation = { showDialog = false }, context, sendData, macroLabels, macroIcons, macroFunctions)
-            }
-            IconButton(
-                onClick = {
-                    showDialog = true
-                },
-                modifier = Modifier,
-            ) {
-                Icon(
-                    painter = painterResource(R.drawable.add_icon), contentDescription = "ajouter une macro"
-                )
-            }
-        }
-    }
-}
-
-//@Preview(showBackground = true)
-@Composable
-fun MainScreenPreview() {
-    val sampleMacroLabels = arrayListOf("Macro 1")
-    val sampleMacroIcons = arrayListOf(R.drawable.copy_icon)
-    val sampleMacroFunctions = arrayListOf({})
-
-    MainScreen(
-        sendData = { data -> /* Mock sendData function */ },
-        onSettingsButtonClicked = { /* Mock settings button action */ },
-        macroLabels = sampleMacroLabels,
-        macroIcons = sampleMacroIcons,
-        macroFunctions = sampleMacroFunctions
-    )
-}
-@Preview(showBackground = true)
-@Composable
-fun ColorPickerTabPreview() {
-    ColorPickerTab(
-        sendData = { colorHex -> /* Mock sendData function with colorHex */ }
-    )
-}
-
-
-@Composable
-fun ColorPickerTab(
-    sendData: (String) -> Unit
-)
-{
-    val controller = rememberColorPickerController()
-    var hexCode by remember { mutableStateOf("") }
-    var textColor by remember { mutableStateOf(Color.Transparent) }
-    val configuration = LocalConfiguration.current
-    val screenWidth = configuration.screenWidthDp.dp
-    val screenHeight = configuration.screenHeightDp.dp
-
-    val smallHeight = screenHeight / 100
-    val medHeight = screenHeight / 50
-    val bigHeight = screenHeight / 20
-
-    val smallWidth = screenWidth / 100
-    val medWidth = screenWidth / 50
-    val bigWidth = screenWidth / 20
-
-    Card (
-        modifier = Modifier.padding(horizontal = bigWidth, vertical = medWidth),
-        shape = MaterialTheme.shapes.extraLarge,
-        colors = CardDefaults.cardColors(
-            containerColor = Color.LightGray.copy(0.6f) // Couleur de fond de la Card
-        )
-    ) {
-
-        Row() {
-            Column(
-                modifier = Modifier
-                    .weight(0.8f),
-                horizontalAlignment = Alignment.CenterHorizontally
-
-            ) {
-                Box(
-                    modifier = Modifier
-                        .padding(0.dp)
-                        .height(bigHeight * 4)
-                ) {
-                    HsvColorPicker(
-                        modifier = Modifier.padding(smallWidth),
-                        controller = controller,
-                        drawOnPosSelected = {
-                            drawColorIndicator(
-                                controller.selectedPoint.value,
-                                controller.selectedColor.value,
-                            )
-                        },
-                        onColorChanged = { colorEnvelope ->
-                            hexCode = colorEnvelope.hexCode
-                            textColor = colorEnvelope.color
-                        },
-                        initialColor = Color.Red,
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(smallWidth)) // Espacement entre les éléments
-
-                AlphaSlider(
-                    modifier = Modifier
-                        .testTag("HSV_AlphaSlider")
-                        .fillMaxWidth(0.8f)
-                        .padding(smallWidth)
-                        .height(bigHeight)
-                        .align(Alignment.CenterHorizontally),
-                    controller = controller,
-                )
-
-                Spacer(modifier = Modifier.height(smallWidth))
-
-                BrightnessSlider(
-                    modifier = Modifier
-                        .fillMaxWidth(0.8f)
-                        .padding(smallWidth)
-                        .height(bigHeight),
-                    controller = controller,
-                )
-            }
-
-            Column(
-                modifier = Modifier
-                    .weight(0.5f),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-
-                Spacer(modifier = Modifier.height(smallWidth))
-
-                Text(
-                    text = "#$hexCode",
-                    color = textColor,
-                    fontSize = 16.sp,
-                    fontWeight = FontWeight.Bold,
-                    modifier = Modifier
-                        .align(Alignment.CenterHorizontally)
-                        .padding(medWidth),
-                )
-
-                Spacer(modifier = Modifier.height(smallWidth))
-
-                AlphaTile(
-                    modifier = Modifier
-                        .size(bigHeight * 3)
-                        .clip(RoundedCornerShape(6.dp))
-                        .padding(smallWidth)
-                        .align(Alignment.CenterHorizontally),
-                    controller = controller,
-                )
-
-                Spacer(modifier = Modifier.height(32.dp))
-
-                Button(onClick = {
-                    sendData("#$hexCode")
-                }) {
-                    Text(
-                        text = "Envoyer",
-                        style = MaterialTheme.typography.labelSmall,
-                        color = Color.White
-                    )
-                }
-
-            }
-        }
-    }
-}
-
-
-@Composable
-fun KeyboardInput(sendData: (String) -> Unit) {
-    var input by remember { mutableStateOf("") }
-    val backspaceKeyCode = Key.Backspace.nativeKeyCode
-    val backspace = """\b"""
-    val focusRequester = remember { FocusRequester() }
-    TextField(
-            value = input,
-            onValueChange = { newText ->
-
-                val addedText =
-                if (newText.isNotEmpty()) {
-                    findFirstDifferenceIndex(newText, input).let {
-                        if (it >= 0) {
-                            for (i in it..input.length-1) {
-                                sendData(backspace)
-                            }
-                            newText.substring(it)
-                        } else {
-                            if(newText.length > input.length){
-                                newText.substring(input.length)
-                            }
-                            else{
-                                for (i in 1..input.length-newText.length) {
-                                    sendData(backspace)
-                                }
-                                ""
-                                }
-                        }
-                    }
-                }else{
-                        sendData(backspace)
-                        ""
-                    }
-
-                input = if (addedText != ".") newText else ""
-
-                if (addedText.isNotEmpty()) {
-                    sendData(addedText)
-                }
-
-            },
-            label = { Text("Label") },
-
-            modifier = Modifier
-                .onKeyEvent { keyEvent ->
-                    if (keyEvent.nativeKeyEvent.keyCode == backspaceKeyCode && input.isEmpty()) {
-                        sendData(backspace)
-                        true
-                    } else {
-                        false
-                    }
-                }
-                .fillMaxWidth()
-                .alpha(0f)
-                .focusRequester(focusRequester),
-
-            singleLine = true,
-            textStyle = TextStyle(color = Color.Transparent), // Le texte est également invisible,
-
-        )
-    LaunchedEffect(Unit) {
-        focusRequester.requestFocus()
-    }
-
-}
-
-
-fun findFirstDifferenceIndex(str1: String, str2: String): Int {
-    return str1.zip(str2).indexOfFirst { (char1, char2) -> char1 != char2 }
-}
-
-@Composable
-fun TouchPad(
-    sendData: (String) -> Unit
-) {
-
-    val context = LocalContext.current
-
-    // PARAMETERS
-    val deadZone = 1
-    val sensibility = 2
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .border(1.dp, Color.DarkGray)
-            .fillMaxHeight(0.6f)
-            .background(Color.LightGray.copy(0.6f)) // FOR DEBUG
-            .pointerInput(Unit) {
-                awaitPointerEventScope {
-                    var prevClickDroit = false //Besoin de ça car quand 2 doigts puis relaché les tailles des changes sont : 1221, cette variable permet d'ignorer le dernier 1
-                    var isHoldSingle = false
-                    var isHoldDouble = false
-
-                    var firstPositionSingle = Offset.Zero // pour un doigt
-                    var lastPositionSingle = Offset.Zero
-                    var firstHoldTimeSingle = 0L
-
-                    var firstPosition1 = Offset.Zero // pour deux doigts
-                    var firstPosition2 = Offset.Zero
-                    var lastPosition1 = Offset.Zero
-                    var lastPosition2 = Offset.Zero
-                    var firstHoldTimeDouble = 0L
-
-
-                    while (true) {
-                        val event = awaitPointerEvent()
-                        if (event.changes.size == 2) {
-                            // Gestion à 2 doigts
-                            val change1 = event.changes[0]
-                            val change2 = event.changes[1]
-
-                            if (!isHoldDouble && change1.pressed && change2.pressed) {
-                                // Premier appui
-                                firstPosition1 = change1.position
-                                firstPosition2 = change2.position
-                                lastPosition1 = change1.position
-                                lastPosition2 = change2.position
-                                firstHoldTimeDouble = change1.uptimeMillis
-                                isHoldDouble = true
-                            } else if (isHoldDouble && (!change1.pressed || !change2.pressed)) {
-                                // Détection clic droit
-                                val delta1 = firstPosition1 - lastPosition1
-                                val delta2 = firstPosition2 - lastPosition2
-                                val deltaTime = abs(change1.uptimeMillis - firstHoldTimeDouble)
-
-                                if (delta1.getDistance() <= deadZone && delta2.getDistance() <= deadZone && deltaTime < 500) {
-                                    // Clic droit
-                                    sendData(context.getString(R.string.id_click_droit))
-                                    change1.consume()
-                                    change2.consume()
-                                    prevClickDroit = true
-                                    lastPositionSingle = if (change1.pressed) change1.position else change2.position
-                                }
-                                isHoldDouble = false
-                            } else if (isHoldDouble && change1.pressed && change2.pressed) {
-                                // Détection défilement
-                                val currentPosition1 = change1.position
-                                val currentPosition2 = change2.position
-                                val offsetDelta = (currentPosition1.y + currentPosition2.y) / 2 -
-                                        (lastPosition1.y + lastPosition2.y) / 2
-
-                                if (abs(offsetDelta) > deadZone) {
-                                    val scrollDirection = if (offsetDelta > 0) -1 else 1
-                                    val scrollCommand =
-                                        context.getString(R.string.id_mouse_scroll) + ":H:$scrollDirection"
-                                    sendData(scrollCommand)
-                                    change1.consume()
-                                    change2.consume()
-                                }
-
-                                // Mettre à jour la dernière position
-                                lastPosition1 = currentPosition1
-                                lastPosition2 = currentPosition2
-                            }
-
-
-                    } else if (event.changes.size == 1) {
-                            val change = event.changes[0]
-                            if (prevClickDroit) {
-                                prevClickDroit = false
-                                lastPositionSingle = change.position
-                            } else {
-                                // Gestion à un doigt
-                                if (!isHoldSingle && change.pressed) {
-                                    firstPositionSingle = change.position
-                                    lastPositionSingle = change.position
-                                    firstHoldTimeSingle = change.uptimeMillis
-                                    isHoldSingle = true
-                                } else if (isHoldSingle && !change.pressed) {
-                                    val delta = firstPositionSingle - lastPositionSingle
-                                    val deltaTime = abs(change.uptimeMillis - firstHoldTimeSingle)
-
-                                    if (delta.getDistance() < deadZone && deltaTime < 500) {
-                                        // Clic gauche
-                                        sendData(context.getString(R.string.id_click_gauche))
-                                        change.consume()
-                                    }
-                                    isHoldSingle = false
-                                } else if (isHoldSingle && change.pressed) {
-                                    val currentPosition = change.position
-                                    val delta = currentPosition - lastPositionSingle
-
-                                    if (abs(delta.x) > deadZone || abs(delta.y) > deadZone) {
-                                        val moveCommand =
-                                            context.getString(R.string.id_mouse_move) +
-                                                    ":${(sensibility * delta.x).toInt()}:${(sensibility * delta.y).toInt()}"
-                                        sendData(moveCommand)
-                                        change.consume()
-                                    }
-
-                                    lastPositionSingle = currentPosition
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-    )
-
-
-
-}
-
-@Composable
-fun AddMacrosDialog(
-    onDismissRequest: () -> Unit,
-    onConfirmation: () -> Unit,
-    context : Context,
-    sendData: (String) -> Unit,
-    macroLabels : ArrayList<String>,
-    macroIcons : ArrayList<Int>,
-    macroFunctions : ArrayList<() -> Unit>
-)
-{
-    var tempMacroLabels by remember { mutableStateOf(ArrayList<String>()) }
-    var tempMacroIcons by remember { mutableStateOf(ArrayList<Int>()) }
-    var tempMacroFunctions by remember { mutableStateOf(ArrayList<() -> Unit>()) }
-
-    tempMacroLabels.addAll(macroLabels)
-    tempMacroIcons.addAll(macroIcons)
-    tempMacroFunctions.addAll(macroFunctions)
-
-    Dialog(onDismissRequest = { onDismissRequest() })
-    {
-        Card(
-            colors = CardDefaults.cardColors(MaterialTheme.colorScheme.background),
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(375.dp)
-                .padding(16.dp),
-            shape = RoundedCornerShape(16.dp),
-        )
-        {
-            Column (modifier = Modifier
-                .align(Alignment.CenterHorizontally)
-                .fillMaxHeight()
-                .padding(top = 20.dp, bottom = 20.dp)
-            ){
-                Text(
-                    text = "Selectionnez les macros à ajouter.",
-                    modifier = Modifier
-                        .align(Alignment.CenterHorizontally)
-                        .padding(bottom = 10.dp)
-                )
-
-                OutlinedCard(
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.surface,
-                    ),
-                    modifier = Modifier
-                        .size(width = 240.dp, height = 200.dp)
-                        .align(Alignment.CenterHorizontally)
-                )
-                {
-                    Column(modifier = Modifier
-                        .align(Alignment.CenterHorizontally)
-                        .verticalScroll(rememberScrollState(),)
-                    ){
-                        MacroChip(
-                            "couper", R.drawable.cut_icon, {sendData(context.getString(R.string.id_cut))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                        MacroChip(
-                            "tout selectionner", R.drawable.select_all_icon, {sendData(context.getString(R.string.id_select_all))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                        MacroChip(
-                            "undo", R.drawable.undo_icon, {sendData(context.getString(R.string.id_undo))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                        MacroChip(
-                            "redo", R.drawable.redo_icon, {sendData(context.getString(R.string.id_redo))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                        MacroChip(
-                            "chercher", R.drawable.find_icon, {sendData(context.getString(R.string.id_search))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                        MacroChip(
-                            "chercher et remplacer", R.drawable.find_replace_icon, {sendData(context.getString(R.string.id_search_replace))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                        MacroChip(
-                            "imprimer", R.drawable.print_icon, {sendData(context.getString(R.string.id_print))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                        MacroChip(
-                            "gras", R.drawable.bold_icon, {sendData(context.getString(R.string.id_gras))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                        MacroChip(
-                            "italique", R.drawable.italic_icon, {sendData(context.getString(R.string.id_italique))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                        MacroChip(
-                            "souligner", R.drawable.underline_icon, {sendData(context.getString(R.string.id_souligne))}, tempMacroLabels, tempMacroIcons, tempMacroFunctions, Modifier.align(Alignment.CenterHorizontally)
-                        )
-                    }
-                }
-
-                Row(modifier = Modifier
-                    .align(Alignment.CenterHorizontally)
-                ){
-                    TextButton(
-                        onClick = {
-                            macroLabels.clear()
-                            macroIcons.clear()
-                            macroFunctions.clear()
-
-                            macroLabels.addAll(tempMacroLabels)
-                            macroIcons.addAll(tempMacroIcons)
-                            macroFunctions.addAll(tempMacroFunctions)
-
-                            onConfirmation()
-                        },
-                        modifier = Modifier.padding(8.dp),
-                    ) {
-                        Text("Ajouter")
-                    }
-                    TextButton(
-                        onClick = {
-                            tempMacroLabels.clear()
-                            tempMacroIcons.clear()
-                            tempMacroFunctions.clear()
-
-                            onDismissRequest()
-                        },
-                        modifier = Modifier.padding(8.dp),
-                    ) {
-                        Text("Annuler")
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun MacroChip(
-    label: String,
-    @DrawableRes iconResource: Int,
-    onClick: () -> Unit,
-    macroLabels : ArrayList<String>,
-    macroIcons : ArrayList<Int>,
-    macroFunctions : ArrayList<() -> Unit>,
-    modifier: Modifier
-) {
-    var selected by remember { mutableStateOf(false) }
-
-    FilterChip(
-        onClick = {
-            selected = !selected
-            if (selected) {
-                if (!macroLabels.contains(label)) {
-                    macroLabels.add(label)
-                    macroIcons.add(iconResource)
-                    macroFunctions.add(onClick)
-                }
-            }
-            else {
-                macroLabels.remove(label)
-                macroIcons.remove(iconResource)
-                macroFunctions.remove(onClick)
-            }
-        },
-        label = {
-            Icon(painter = painterResource(iconResource), contentDescription = label)
-            Text(label)
-        },
-        selected = selected,
-    )
-}
-
-@Composable
-fun GridItem(
-    onClick: () -> Unit,
-    painter: Int,
-    contentDescription: String
-)
-{
-    IconButton(
-        onClick = onClick,
-    ) {
-        Icon(
-            painter = painterResource(painter),
-            contentDescription = contentDescription
-        )
-    }
-}
