+++ app/src/main/java/com/example/klavier/ui/components/KeyboardInput.kt
@@ -1,0 +1,107 @@
+package com.example.klavier.ui.components
+
+import android.content.Context
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.key.Key
+import androidx.compose.ui.input.key.nativeKeyCode
+import androidx.compose.ui.input.key.onKeyEvent
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.input.ImeAction
+import com.example.klavier.R
+
+@Composable
+fun KeyboardInput(sendData: (String) -> Unit, context : Context,) {
+    var input by remember { mutableStateOf("") }
+    val backspaceKeyCode = Key.Backspace.nativeKeyCode
+    val backspace = """\b"""
+    val focusRequester = remember { FocusRequester() }
+    TextField(
+        value = input,
+        onValueChange = { newText ->
+
+            val addedText =
+                if (newText.isNotEmpty()) {
+                    findFirstDifferenceIndex(newText, input).let {
+                        if (it >= 0) {
+                            for (i in it..input.length-1) {
+                                sendData(backspace)
+                            }
+                            newText.substring(it)
+                        } else {
+                            if(newText.length > input.length){
+                                newText.substring(input.length)
+                            }
+                            else{
+                                for (i in 1..input.length-newText.length) {
+                                    sendData(backspace)
+                                }
+                                ""
+                            }
+                        }
+                    }
+                }else{
+                    sendData(backspace)
+                    ""
+                }
+
+            input = if (addedText != ".") newText else ""
+
+            if (addedText.isNotEmpty()) {
+                sendData(addedText)
+            }
+
+        },
+        label = { Text("Label") },
+
+        modifier = Modifier
+            .onKeyEvent { keyEvent ->
+                if (keyEvent.nativeKeyEvent.keyCode == backspaceKeyCode && input.isEmpty()) {
+                    sendData(backspace)
+                    true
+                } else {
+                    false
+                }
+            }
+            .fillMaxWidth()
+            .alpha(0f)
+            .focusRequester(focusRequester),
+
+        singleLine = true,
+        keyboardOptions = KeyboardOptions.Default.copy(
+            imeAction = ImeAction.Done // Action Key
+        ),
+        keyboardActions = KeyboardActions(
+            onDone = {
+                input = ""
+                sendData(context.getString(R.string.id_enter))
+            }
+        ),
+        textStyle = TextStyle(color = Color.Transparent), // Le texte est également invisible,
+
+
+    )
+    LaunchedEffect(Unit) {
+        focusRequester.requestFocus()
+    }
+
+}
+
+
+fun findFirstDifferenceIndex(str1: String, str2: String): Int {
+    return str1.zip(str2).indexOfFirst { (char1, char2) -> char1 != char2 }
+}
