+++ app/src/main/java/com/example/klavier/ui/screens/SplashScreen.kt
@@ -1,4 +1,4 @@
-package com.example.klavier.ui
+package com.example.klavier.ui.screens
 
 import android.content.Context
 import androidx.compose.animation.AnimatedVisibility
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.Button
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.material3.ButtonDefaults.buttonColors
@@ -23,9 +22,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.example.klavier.R
 import kotlin.reflect.KFunction1
 import androidx.compose.ui.platform.LocalContext
