+++ app/src/main/res/values/strings.xml
@@ -23,4 +23,6 @@
     <string name="id_mouse_move">#$M1</string>
     <string name="id_mouse_scroll">#$M2</string>
 
+    <string name="id_enter">#$E</string>
+
 </resources>
